refactor(block-runtime): split parser and schema helpers
