+++ .sampo/changesets/issue-339-block-runtime-boundaries-phase-2.md
@@ -1,0 +1,5 @@
+---
+npm/@wp-typia/block-runtime: patch
+---
+
+Split schema-core and metadata-parser helper clusters into dedicated internal modules while preserving the published block-runtime API surface.
