+++ packages/wp-typia-block-runtime/src/metadata-parser.ts
@@ -1,4 +1,3 @@
-import * as fs from "node:fs";
 import * as path from "node:path";
 
 import ts from "typescript";
@@ -7,35 +6,26 @@ import {
 	type AnalysisContext,
 	createAnalysisContext,
 } from "./metadata-analysis.js";
-import { cloneJsonValue } from "./json-utils.js";
 import {
-	type AttributeConstraints,
 	type AttributeNode,
-	type JsonValue,
-	type WordPressAttributeSource,
 	baseNode,
 	cloneProperties,
 	defaultAttributeConstraints,
 	withRequired,
 } from "./metadata-model.js";
-
-const SUPPORTED_TAGS = new Set([
-	"Default",
-	"ExclusiveMaximum",
-	"ExclusiveMinimum",
-	"Format",
-	"MaxLength",
-	"MaxItems",
-	"Maximum",
-	"MinLength",
-	"MinItems",
-	"Minimum",
-	"MultipleOf",
-	"Pattern",
-	"Selector",
-	"Source",
-	"Type",
-]);
+import {
+	getReferenceName,
+	isSerializableExternalDeclaration,
+	resolveIndexedAccessPropertyDeclaration,
+	resolveSymbol,
+} from "./metadata-parser-symbols.js";
+import {
+	applyTag,
+	extractLiteralValue,
+	getPropertyName,
+	getSupportedTagName,
+	mergePrimitiveIntersection,
+} from "./metadata-parser-tags.js";
 
 /**
  * Analyze one named source type from a TypeScript module.
@@ -634,603 +624,4 @@ function parseTypeReference(
 	}
 
 	return parseNamedDeclaration(declaration, ctx, pathLabel, true);
-}
-
-function resolveIndexedAccessPropertyDeclaration(
-	objectTypeNode: ts.TypeNode,
-	propertyKey: string,
-	ctx: AnalysisContext,
-	pathLabel: string,
-): ts.PropertySignature | ts.PropertyDeclaration {
-	const objectType = ctx.checker.getTypeFromTypeNode(objectTypeNode);
-	const propertySymbol = ctx.checker
-		.getPropertiesOfType(objectType)
-		.find((candidate) => candidate.name === propertyKey);
-
-	if (propertySymbol === undefined) {
-		throw new Error(
-			`Indexed access could not resolve property "${propertyKey}" at ${pathLabel}`,
-		);
-	}
-
-	const valueDeclaration = propertySymbol.valueDeclaration;
-	const declaration =
-		valueDeclaration !== undefined &&
-		(ts.isPropertySignature(valueDeclaration) ||
-			ts.isPropertyDeclaration(valueDeclaration))
-			? valueDeclaration
-			: propertySymbol.declarations?.find(
-					(
-						candidate,
-					): candidate is ts.PropertySignature | ts.PropertyDeclaration =>
-						ts.isPropertySignature(candidate) ||
-						ts.isPropertyDeclaration(candidate),
-			  );
-	if (declaration === undefined) {
-		throw new Error(
-			`Indexed access property "${propertyKey}" does not resolve to a typed property at ${pathLabel}`,
-		);
-	}
-	if (!isSerializableExternalDeclaration(declaration, ctx)) {
-		throw new Error(
-			`External or non-serializable indexed access property "${propertyKey}" is not supported at ${pathLabel}`,
-		);
-	}
-
-	return declaration;
-}
-
-function mergePrimitiveIntersection(
-	nodes: AttributeNode[],
-	pathLabel: string,
-): AttributeNode {
-	const [first, ...rest] = nodes;
-
-	if (!isPrimitiveCompatibleNode(first)) {
-		throw new Error(
-			`Unsupported intersection at ${pathLabel}; only primitive-compatible intersections are supported`,
-		);
-	}
-
-	let merged = withRequired(first, first.required);
-	for (const node of rest) {
-		if (!isPrimitiveCompatibleNode(node) || node.kind !== merged.kind) {
-			throw new Error(
-				`Unsupported intersection at ${pathLabel}; only a single primitive kind plus typia tags is supported`,
-			);
-		}
-
-		merged = {
-			...merged,
-			constraints: mergeConstraints(
-				merged.constraints,
-				node.constraints,
-				pathLabel,
-			),
-			defaultValue: mergeDefaultValue(
-				merged.defaultValue,
-				node.defaultValue,
-				pathLabel,
-			),
-			enumValues: intersectEnumValues(
-				merged.enumValues,
-				node.enumValues,
-				pathLabel,
-			),
-			required: merged.required && node.required,
-		};
-	}
-
-	return merged;
-}
-
-function isPrimitiveCompatibleNode(node: AttributeNode): boolean {
-	return (
-		(node.kind === "string" ||
-			node.kind === "number" ||
-			node.kind === "boolean") &&
-		node.items === undefined &&
-		node.properties === undefined &&
-		node.union === null
-	);
-}
-
-function mergeConstraints(
-	left: AttributeConstraints,
-	right: AttributeConstraints,
-	pathLabel: string,
-): AttributeConstraints {
-	return {
-		exclusiveMaximum: mergeMaximumLike(
-			left.exclusiveMaximum,
-			right.exclusiveMaximum,
-		),
-		exclusiveMinimum: mergeMinimumLike(
-			left.exclusiveMinimum,
-			right.exclusiveMinimum,
-		),
-		format: mergeExactLike(left.format, right.format, pathLabel, "format"),
-		maxLength: mergeMaximumLike(left.maxLength, right.maxLength),
-		maxItems: mergeMaximumLike(left.maxItems, right.maxItems),
-		maximum: mergeMaximumLike(left.maximum, right.maximum),
-		minLength: mergeMinimumLike(left.minLength, right.minLength),
-		minItems: mergeMinimumLike(left.minItems, right.minItems),
-		minimum: mergeMinimumLike(left.minimum, right.minimum),
-		multipleOf: mergeExactLike(
-			left.multipleOf,
-			right.multipleOf,
-			pathLabel,
-			"multipleOf",
-		),
-		pattern: mergeExactLike(left.pattern, right.pattern, pathLabel, "pattern"),
-		typeTag: mergeExactLike(left.typeTag, right.typeTag, pathLabel, "typeTag"),
-	};
-}
-
-function mergeMinimumLike(
-	left: number | null,
-	right: number | null,
-): number | null {
-	if (left === null) {
-		return right;
-	}
-	if (right === null) {
-		return left;
-	}
-	return Math.max(left, right);
-}
-
-function mergeMaximumLike(
-	left: number | null,
-	right: number | null,
-): number | null {
-	if (left === null) {
-		return right;
-	}
-	if (right === null) {
-		return left;
-	}
-	return Math.min(left, right);
-}
-
-function mergeExactLike<T extends string | number>(
-	left: T | null,
-	right: T | null,
-	pathLabel: string,
-	label: string,
-): T | null {
-	if (left === null) {
-		return right;
-	}
-	if (right === null) {
-		return left;
-	}
-	if (left !== right) {
-		throw new Error(
-			`Conflicting ${label} constraints in intersection at ${pathLabel}: ${left} vs ${right}`,
-		);
-	}
-	return left;
-}
-
-function mergeDefaultValue(
-	left: JsonValue | undefined,
-	right: JsonValue | undefined,
-	pathLabel: string,
-): JsonValue | undefined {
-	if (left === undefined) {
-		return right;
-	}
-	if (right === undefined) {
-		return left;
-	}
-	if (!jsonValuesEqual(left, right)) {
-		throw new Error(
-			`Conflicting default values in intersection at ${pathLabel}`,
-		);
-	}
-	return cloneJsonValue(left);
-}
-
-function jsonValuesEqual(left: JsonValue, right: JsonValue): boolean {
-	if (left === right) {
-		return true;
-	}
-	if (
-		left === null ||
-		right === null ||
-		typeof left !== "object" ||
-		typeof right !== "object"
-	) {
-		return false;
-	}
-	if (Array.isArray(left) || Array.isArray(right)) {
-		if (!Array.isArray(left) || !Array.isArray(right)) {
-			return false;
-		}
-		if (left.length !== right.length) {
-			return false;
-		}
-
-		return left.every((value, index) => jsonValuesEqual(value, right[index]));
-	}
-
-	const leftEntries = Object.entries(left);
-	const rightEntries = Object.entries(right);
-	if (leftEntries.length !== rightEntries.length) {
-		return false;
-	}
-
-	return leftEntries.every(([key, value]) =>
-		Object.prototype.hasOwnProperty.call(right, key) &&
-		jsonValuesEqual(value, right[key]),
-	);
-}
-
-function intersectEnumValues(
-	left: Array<string | number | boolean> | null,
-	right: Array<string | number | boolean> | null,
-	pathLabel: string,
-): Array<string | number | boolean> | null {
-	if (left === null) {
-		return right ? [...right] : null;
-	}
-	if (right === null) {
-		return [...left];
-	}
-
-	const allowed = new Set(right);
-	const intersection = left.filter((value) => allowed.has(value));
-	if (intersection.length === 0) {
-		throw new Error(`Intersection at ${pathLabel} resolves to an empty enum`);
-	}
-	return intersection;
-}
-
-function applyTag(
-	node: AttributeNode,
-	tagNode: ts.TypeReferenceNode,
-	pathLabel: string,
-): void {
-	const tagName = getSupportedTagName(tagNode);
-	if (tagName === null) {
-		return;
-	}
-
-	const [arg] = tagNode.typeArguments ?? [];
-	if (arg === undefined) {
-		throw new Error(
-			`Tag "${tagName}" is missing its generic argument at ${pathLabel}`,
-		);
-	}
-
-	switch (tagName) {
-		case "Default": {
-			const value = parseDefaultValue(arg, pathLabel);
-			if (value === undefined) {
-				throw new Error(
-					`Unsupported Default value at ${pathLabel}: ${arg.getText()}`,
-				);
-			}
-			node.defaultValue = value;
-			return;
-		}
-		case "Format":
-			node.constraints.format = parseStringLikeArgument(
-				arg,
-				tagName,
-				pathLabel,
-			);
-			return;
-		case "Pattern":
-			node.constraints.pattern = parseStringLikeArgument(
-				arg,
-				tagName,
-				pathLabel,
-			);
-			return;
-		case "Selector":
-			node.wp.selector = parseStringLikeArgument(arg, tagName, pathLabel);
-			return;
-		case "Source":
-			node.wp.source = parseWordPressAttributeSource(arg, pathLabel);
-			return;
-		case "Type":
-			node.constraints.typeTag = parseStringLikeArgument(
-				arg,
-				tagName,
-				pathLabel,
-			);
-			return;
-		case "MinLength":
-			node.constraints.minLength = parseNumericArgument(
-				arg,
-				tagName,
-				pathLabel,
-			);
-			return;
-		case "MaxLength":
-			node.constraints.maxLength = parseNumericArgument(
-				arg,
-				tagName,
-				pathLabel,
-			);
-			return;
-		case "MinItems":
-			node.constraints.minItems = parseNumericArgument(arg, tagName, pathLabel);
-			return;
-		case "MaxItems":
-			node.constraints.maxItems = parseNumericArgument(arg, tagName, pathLabel);
-			return;
-		case "Minimum":
-			node.constraints.minimum = parseNumericArgument(arg, tagName, pathLabel);
-			return;
-		case "Maximum":
-			node.constraints.maximum = parseNumericArgument(arg, tagName, pathLabel);
-			return;
-		case "ExclusiveMinimum":
-			node.constraints.exclusiveMinimum = parseNumericArgument(
-				arg,
-				tagName,
-				pathLabel,
-			);
-			return;
-		case "ExclusiveMaximum":
-			node.constraints.exclusiveMaximum = parseNumericArgument(
-				arg,
-				tagName,
-				pathLabel,
-			);
-			return;
-		case "MultipleOf":
-			node.constraints.multipleOf = parseNumericArgument(
-				arg,
-				tagName,
-				pathLabel,
-			);
-			return;
-		default:
-			return;
-	}
-}
-
-function parseDefaultValue(
-	node: ts.TypeNode,
-	pathLabel: string,
-): JsonValue | undefined {
-	if (ts.isParenthesizedTypeNode(node)) {
-		return parseDefaultValue(node.type, pathLabel);
-	}
-	if (ts.isLiteralTypeNode(node)) {
-		const literal = extractLiteralValue(node);
-		return literal === undefined ? undefined : literal;
-	}
-	if (ts.isTypeLiteralNode(node)) {
-		const objectValue: Record<string, JsonValue> = {};
-		for (const member of node.members) {
-			if (!ts.isPropertySignature(member) || member.type === undefined) {
-				throw new Error(`Unsupported object Default value at ${pathLabel}`);
-			}
-			const propertyName = getPropertyName(member.name);
-			const value = parseDefaultValue(
-				member.type,
-				`${pathLabel}.${propertyName}`,
-			);
-			if (value === undefined) {
-				throw new Error(
-					`Unsupported object Default value at ${pathLabel}.${propertyName}`,
-				);
-			}
-			objectValue[propertyName] = value;
-		}
-		return objectValue;
-	}
-	if (ts.isTupleTypeNode(node)) {
-		return node.elements.map((element, index) => {
-			const value = parseDefaultValue(element, `${pathLabel}[${index}]`);
-			if (value === undefined) {
-				throw new Error(
-					`Unsupported array Default value at ${pathLabel}[${index}]`,
-				);
-			}
-			return value;
-		});
-	}
-	if (node.kind === ts.SyntaxKind.NullKeyword) {
-		return null;
-	}
-	return undefined;
-}
-
-function parseNumericArgument(
-	node: ts.TypeNode,
-	tagName: string,
-	pathLabel: string,
-): number {
-	const value = extractLiteralValue(node);
-	if (typeof value !== "number") {
-		throw new Error(
-			`Tag "${tagName}" expects a numeric literal at ${pathLabel}`,
-		);
-	}
-	return value;
-}
-
-function parseStringLikeArgument(
-	node: ts.TypeNode,
-	tagName: string,
-	pathLabel: string,
-): string {
-	const value = extractLiteralValue(node);
-	if (typeof value !== "string") {
-		throw new Error(
-			`Tag "${tagName}" expects a string literal at ${pathLabel}`,
-		);
-	}
-	return value;
-}
-
-function parseWordPressAttributeSource(
-	node: ts.TypeNode,
-	pathLabel: string,
-): WordPressAttributeSource {
-	const value = parseStringLikeArgument(node, "Source", pathLabel);
-	if (value === "html" || value === "text" || value === "rich-text") {
-		return value;
-	}
-
-	throw new Error(
-		`Tag "Source" only supports "html", "text", or "rich-text" at ${pathLabel}`,
-	);
-}
-
-function extractLiteralValue(
-	node: ts.TypeNode | ts.Node,
-): string | number | boolean | undefined {
-	if (ts.isParenthesizedTypeNode(node)) {
-		return extractLiteralValue(node.type);
-	}
-	if (ts.isLiteralTypeNode(node)) {
-		return extractLiteralValue(node.literal);
-	}
-	if (
-		ts.isPrefixUnaryExpression(node) &&
-		node.operator === ts.SyntaxKind.MinusToken &&
-		ts.isNumericLiteral(node.operand)
-	) {
-		return -Number(node.operand.text);
-	}
-	if (node.kind === ts.SyntaxKind.TrueKeyword) {
-		return true;
-	}
-	if (node.kind === ts.SyntaxKind.FalseKeyword) {
-		return false;
-	}
-	if (ts.isStringLiteral(node)) {
-		return node.text;
-	}
-	if (ts.isNumericLiteral(node)) {
-		return Number(node.text);
-	}
-	return undefined;
-}
-
-function getPropertyName(name: ts.PropertyName): string {
-	if (
-		ts.isIdentifier(name) ||
-		ts.isStringLiteral(name) ||
-		ts.isNumericLiteral(name)
-	) {
-		return name.text;
-	}
-	throw new Error(`Unsupported property name: ${name.getText()}`);
-}
-
-function getSupportedTagName(node: ts.TypeReferenceNode): string | null {
-	const typeName = getEntityNameText(node.typeName);
-	const [, tagName] = typeName.split(".");
-	if (
-		!typeName.startsWith("tags.") ||
-		tagName === undefined ||
-		!SUPPORTED_TAGS.has(tagName)
-	) {
-		return null;
-	}
-	return tagName;
-}
-
-function getEntityNameText(name: ts.EntityName): string {
-	if (ts.isIdentifier(name)) {
-		return name.text;
-	}
-	return `${getEntityNameText(name.left)}.${name.right.text}`;
-}
-
-function resolveSymbol(
-	node: ts.TypeReferenceNode | ts.ExpressionWithTypeArguments,
-	checker: ts.TypeChecker,
-): ts.Symbol | undefined {
-	const symbol = checker.getSymbolAtLocation(
-		ts.isTypeReferenceNode(node) ? node.typeName : node.expression,
-	);
-	if (symbol === undefined) {
-		return undefined;
-	}
-	return symbol.flags & ts.SymbolFlags.Alias
-		? checker.getAliasedSymbol(symbol)
-		: symbol;
-}
-
-function getReferenceName(
-	node: ts.TypeReferenceNode | ts.ExpressionWithTypeArguments,
-): string {
-	if (ts.isTypeReferenceNode(node)) {
-		return getEntityNameText(node.typeName);
-	}
-	return node.expression.getText();
-}
-
-function isProjectLocalDeclaration(
-	declaration: ts.Declaration,
-	projectRoot: string,
-): boolean {
-	const fileName = declaration.getSourceFile().fileName;
-	return (
-		!fileName.includes("node_modules") &&
-		!path.relative(projectRoot, fileName).startsWith("..")
-	);
-}
-
-function isSerializableExternalDeclaration(
-	declaration: ts.Declaration,
-	ctx: AnalysisContext,
-): boolean {
-	if (isProjectLocalDeclaration(declaration, ctx.projectRoot)) {
-		return true;
-	}
-
-	const packageName = getOwningPackageName(
-		declaration.getSourceFile().fileName,
-		ctx.packageNameCache,
-	);
-	return packageName !== null && ctx.allowedExternalPackages.has(packageName);
-}
-
-function getOwningPackageName(
-	fileName: string,
-	cache: Map<string, string | null>,
-): string | null {
-	let currentDir = path.dirname(fileName);
-
-	while (true) {
-		if (cache.has(currentDir)) {
-			return cache.get(currentDir) ?? null;
-		}
-
-		const packageJsonPath = path.join(currentDir, "package.json");
-		if (fs.existsSync(packageJsonPath)) {
-			try {
-				const packageJson = JSON.parse(
-					fs.readFileSync(packageJsonPath, "utf8"),
-				) as {
-					name?: string;
-				};
-				const packageName =
-					typeof packageJson.name === "string" ? packageJson.name : null;
-				cache.set(currentDir, packageName);
-				return packageName;
-			} catch {
-				cache.set(currentDir, null);
-				return null;
-			}
-		}
-
-		const parentDir = path.dirname(currentDir);
-		if (parentDir === currentDir) {
-			cache.set(currentDir, null);
-			return null;
-		}
-
-		currentDir = parentDir;
-	}
 }
