+++ packages/wp-typia-block-runtime/src/schema-core.ts
@@ -5,6 +5,15 @@ import type {
 	ManifestDocument,
 	ManifestUnionMetadata,
 } from "./migration-types.js";
+import {
+	createBootstrapResponseHeaders,
+	normalizeEndpointAuthDefinition,
+} from "./schema-core-auth.js";
+import {
+	projectSchemaObjectForAiStructuredOutput,
+	projectSchemaObjectForRest,
+} from "./schema-core-projection.js";
+export { normalizeEndpointAuthDefinition } from "./schema-core-auth.js";
 
 export interface JsonSchemaObject {
 	[key: string]: JsonValue | JsonSchemaObject | JsonSchemaObject[] | undefined;
@@ -270,10 +279,6 @@ const WP_TYPIA_OPENAPI_LITERALS = {
 	WP_REST_NONCE_SCHEME: "wpRestNonce",
 } as const;
 
-const WP_TYPIA_SCHEMA_UINT32_MAX = 4_294_967_295;
-const WP_TYPIA_SCHEMA_INT32_MAX = 2_147_483_647;
-const WP_TYPIA_SCHEMA_INT32_MIN = -2_147_483_648;
-
 function applyConstraintIfNumber(
 	schema: JsonSchemaObject,
 	key: string,
@@ -345,382 +350,6 @@ function addDiscriminatorToObjectBranch(
 		: [discriminator];
 	schema.required = required;
 	return schema;
-}
-
-function isJsonSchemaObject(value: unknown): value is JsonSchemaObject {
-	return value !== null && typeof value === "object" && !Array.isArray(value);
-}
-
-function cloneJsonSchemaNode<T extends JsonValue | JsonSchemaObject>(value: T): T {
-	if (Array.isArray(value)) {
-		return value.map((item) =>
-			isJsonSchemaObject(item) || Array.isArray(item)
-				? cloneJsonSchemaNode(item as JsonValue | JsonSchemaObject)
-				: item,
-		) as T;
-	}
-
-	if (!isJsonSchemaObject(value)) {
-		return value;
-	}
-
-	return Object.fromEntries(
-		Object.entries(value).map(([key, child]) => [
-			key,
-			child === undefined
-				? undefined
-				: isJsonSchemaObject(child) || Array.isArray(child)
-					? cloneJsonSchemaNode(child as JsonValue | JsonSchemaObject)
-					: child,
-		]),
-	) as T;
-}
-
-function isWpTypiaSchemaExtensionKey(key: string): boolean {
-	return key.startsWith("x-wp-typia-");
-}
-
-function getProjectedNumericMultipleOf(
-	schema: JsonSchemaObject,
-	path: string,
-	typeTag: string,
-): number {
-	const currentMultipleOf = schema.multipleOf;
-	if (
-		currentMultipleOf !== undefined &&
-		(typeof currentMultipleOf !== "number" ||
-			!Number.isFinite(currentMultipleOf) ||
-			!Number.isInteger(currentMultipleOf) ||
-			currentMultipleOf <= 0)
-	) {
-		throw new Error(
-			`Unable to project unsupported ${typeTag} multipleOf at "${path}".`,
-		);
-	}
-
-	return typeof currentMultipleOf === "number" ? currentMultipleOf : 1;
-}
-
-function applyProjectedUint32Constraints(
-	schema: JsonSchemaObject,
-	path: string,
-): void {
-	const currentMinimum = schema.minimum;
-	if (typeof currentMinimum === "number" && Number.isFinite(currentMinimum)) {
-		schema.minimum = Math.max(currentMinimum, 0);
-	} else {
-		schema.minimum = 0;
-	}
-
-	const currentMaximum = schema.maximum;
-	if (typeof currentMaximum === "number" && Number.isFinite(currentMaximum)) {
-		schema.maximum = Math.min(currentMaximum, WP_TYPIA_SCHEMA_UINT32_MAX);
-	} else {
-		schema.maximum = WP_TYPIA_SCHEMA_UINT32_MAX;
-	}
-
-	schema.multipleOf = getProjectedNumericMultipleOf(schema, path, "uint32");
-	schema.type = "integer";
-}
-
-function applyProjectedInt32Constraints(
-	schema: JsonSchemaObject,
-	path: string,
-): void {
-	const currentMinimum = schema.minimum;
-	if (typeof currentMinimum === "number" && Number.isFinite(currentMinimum)) {
-		schema.minimum = Math.max(currentMinimum, WP_TYPIA_SCHEMA_INT32_MIN);
-	} else {
-		schema.minimum = WP_TYPIA_SCHEMA_INT32_MIN;
-	}
-
-	const currentMaximum = schema.maximum;
-	if (typeof currentMaximum === "number" && Number.isFinite(currentMaximum)) {
-		schema.maximum = Math.min(currentMaximum, WP_TYPIA_SCHEMA_INT32_MAX);
-	} else {
-		schema.maximum = WP_TYPIA_SCHEMA_INT32_MAX;
-	}
-
-	schema.multipleOf = getProjectedNumericMultipleOf(schema, path, "int32");
-	schema.type = "integer";
-}
-
-function applyProjectedTypeTag(
-	schema: JsonSchemaObject,
-	typeTag: string,
-	path: string,
-): void {
-	switch (typeTag) {
-		case "uint32":
-			applyProjectedUint32Constraints(schema, path);
-			break;
-		case "int32":
-			applyProjectedInt32Constraints(schema, path);
-			break;
-		case "float":
-		case "double":
-			schema.type = "number";
-			break;
-		default:
-			throw new Error(
-				`Unsupported wp-typia schema type tag "${typeTag}" at "${path}".`,
-			);
-	}
-}
-
-function canProjectTypeTag(typeTag: string): boolean {
-	switch (typeTag) {
-		case "uint32":
-		case "int32":
-		case "float":
-		case "double":
-			return true;
-		default:
-			return false;
-	}
-}
-
-function projectSchemaArrayItemsForAiStructuredOutput(
-	items: JsonSchemaObject[],
-	path: string,
-): JsonSchemaObject[] {
-	return items.map((item, index) =>
-		projectSchemaObjectForAiStructuredOutput(item, `${path}/${index}`),
-	);
-}
-
-function projectSchemaArrayItemsForRest(
-	items: JsonSchemaObject[],
-	path: string,
-): JsonSchemaObject[] {
-	return items.map((item, index) =>
-		projectSchemaObjectForRest(item, `${path}/${index}`),
-	);
-}
-
-function projectSchemaPropertyMapForAiStructuredOutput(
-	properties: Record<string, JsonSchemaObject>,
-	path: string,
-): Record<string, JsonSchemaObject> {
-	return Object.fromEntries(
-		Object.entries(properties).map(([key, value]) => [
-			key,
-			projectSchemaObjectForAiStructuredOutput(value, `${path}/${key}`),
-		]),
-	);
-}
-
-function projectSchemaPropertyMapForRest(
-	properties: Record<string, JsonSchemaObject>,
-	path: string,
-): Record<string, JsonSchemaObject> {
-	return Object.fromEntries(
-		Object.entries(properties).map(([key, value]) => [
-			key,
-			projectSchemaObjectForRest(value, `${path}/${key}`),
-		]),
-	);
-}
-
-function projectSchemaObjectForAiStructuredOutput(
-	node: JsonSchemaObject,
-	path: string,
-): JsonSchemaObject {
-	const projectedNode = cloneJsonSchemaNode(node);
-	const rawTypeTag = projectedNode[WP_TYPIA_OPENAPI_EXTENSION_KEYS.TYPE_TAG];
-
-	if (typeof rawTypeTag === "string") {
-		applyProjectedTypeTag(projectedNode, rawTypeTag, path);
-	}
-
-	delete projectedNode[WP_TYPIA_OPENAPI_EXTENSION_KEYS.TYPE_TAG];
-
-	for (const key of Object.keys(projectedNode)) {
-		if (isWpTypiaSchemaExtensionKey(key)) {
-			delete projectedNode[key];
-			continue;
-		}
-
-		const child = projectedNode[key];
-		if (Array.isArray(child)) {
-			projectedNode[key] = child.every(isJsonSchemaObject)
-				? projectSchemaArrayItemsForAiStructuredOutput(
-						child as JsonSchemaObject[],
-						`${path}/${key}`,
-					)
-				: child;
-			continue;
-		}
-
-		if (!isJsonSchemaObject(child)) {
-			continue;
-		}
-
-		if (key === "properties") {
-			projectedNode[key] = projectSchemaPropertyMapForAiStructuredOutput(
-				child as Record<string, JsonSchemaObject>,
-				`${path}/${key}`,
-			);
-			continue;
-		}
-
-		projectedNode[key] = projectSchemaObjectForAiStructuredOutput(
-			child,
-			`${path}/${key}`,
-		);
-	}
-
-	return projectedNode;
-}
-
-function projectSchemaObjectForRest(
-	node: JsonSchemaObject,
-	path: string,
-): JsonSchemaObject {
-	const projectedNode = cloneJsonSchemaNode(node);
-	const rawTypeTag = projectedNode[WP_TYPIA_OPENAPI_EXTENSION_KEYS.TYPE_TAG];
-
-	if (typeof rawTypeTag === "string" && canProjectTypeTag(rawTypeTag)) {
-		applyProjectedTypeTag(projectedNode, rawTypeTag, path);
-	}
-
-	for (const key of Object.keys(projectedNode)) {
-		const child = projectedNode[key];
-		if (Array.isArray(child)) {
-			projectedNode[key] = child.every(isJsonSchemaObject)
-				? projectSchemaArrayItemsForRest(
-						child as JsonSchemaObject[],
-						`${path}/${key}`,
-					)
-				: child;
-			continue;
-		}
-
-		if (!isJsonSchemaObject(child)) {
-			continue;
-		}
-
-		if (key === "properties") {
-			projectedNode[key] = projectSchemaPropertyMapForRest(
-				child as Record<string, JsonSchemaObject>,
-				`${path}/${key}`,
-			);
-			continue;
-		}
-
-		projectedNode[key] = projectSchemaObjectForRest(child, `${path}/${key}`);
-	}
-
-	applyProjectedBootstrapContract(projectedNode);
-
-	return projectedNode;
-}
-
-function applyProjectedBootstrapContract(schema: JsonSchemaObject): void {
-	if (schema.type !== "object" || !isJsonSchemaObject(schema.properties)) {
-		return;
-	}
-
-	const properties = schema.properties as Record<string, JsonSchemaObject>;
-	if (properties.canWrite?.type !== "boolean") {
-		return;
-	}
-
-	const allOf = Array.isArray(schema.allOf)
-		? [...(schema.allOf as JsonSchemaObject[])]
-		: [];
-	const canWriteIsTrue: JsonSchemaObject = {
-		properties: {
-			canWrite: {
-				const: true,
-			},
-		},
-		required: ["canWrite"],
-	};
-	const buildRequiredPropertyObject = (
-		requiredKeys: string[],
-	): Record<string, JsonSchemaObject> =>
-		Object.fromEntries(
-			requiredKeys.map((requiredKey) => [requiredKey, properties[requiredKey] ?? {}]),
-		);
-	const hasRestNonce = properties.restNonce?.type === "string";
-	const hasPublicWriteCredential =
-		properties.publicWriteToken?.type === "string" &&
-		isJsonSchemaObject(properties.publicWriteExpiresAt);
-
-	if (hasRestNonce && hasPublicWriteCredential) {
-		allOf.push({
-			if: canWriteIsTrue,
-			then: {
-				anyOf: [
-					{
-						properties: buildRequiredPropertyObject(["restNonce"]),
-						required: ["restNonce"],
-					},
-					{
-						properties: buildRequiredPropertyObject([
-							"publicWriteExpiresAt",
-							"publicWriteToken",
-						]),
-						required: ["publicWriteExpiresAt", "publicWriteToken"],
-					},
-				],
-			},
-			else: {
-				not: {
-					anyOf: [
-						{
-							properties: buildRequiredPropertyObject(["restNonce"]),
-							required: ["restNonce"],
-						},
-						{
-							properties: buildRequiredPropertyObject(["publicWriteToken"]),
-							required: ["publicWriteToken"],
-						},
-					],
-				},
-			},
-		});
-	}
-
-	if (hasRestNonce && !hasPublicWriteCredential) {
-		allOf.push({
-			if: canWriteIsTrue,
-			then: {
-				properties: buildRequiredPropertyObject(["restNonce"]),
-				required: ["restNonce"],
-			},
-			else: {
-				not: {
-					properties: buildRequiredPropertyObject(["restNonce"]),
-					required: ["restNonce"],
-				},
-			},
-		});
-	}
-
-	if (hasPublicWriteCredential && !hasRestNonce) {
-		allOf.push({
-			if: canWriteIsTrue,
-			then: {
-				properties: buildRequiredPropertyObject([
-					"publicWriteExpiresAt",
-					"publicWriteToken",
-				]),
-				required: ["publicWriteExpiresAt", "publicWriteToken"],
-			},
-			else: {
-				not: {
-					properties: buildRequiredPropertyObject(["publicWriteToken"]),
-					required: ["publicWriteToken"],
-				},
-			},
-		});
-	}
-
-	if (allOf.length > 0) {
-		schema.allOf = allOf;
-	}
 }
 
 function manifestUnionToJsonSchema(union: ManifestUnionMetadata): JsonSchemaObject {
@@ -919,192 +548,6 @@ function formatEndpointDescription(endpoint: EndpointOpenApiEndpointDefinition):
 	return `${endpoint.operationId} (${endpoint.method} ${endpoint.path})`;
 }
 
-function normalizeWordPressAuthDefinition(
-	wordpressAuth: EndpointWordPressAuthDefinition | undefined,
-): EndpointWordPressAuthDefinition | undefined {
-	if (!wordpressAuth) {
-		return undefined;
-	}
-
-	if (
-		wordpressAuth.mechanism !==
-		WP_TYPIA_OPENAPI_LITERALS.WORDPRESS_PUBLIC_TOKEN_MECHANISM
-	) {
-		return {
-			mechanism: wordpressAuth.mechanism,
-		};
-	}
-
-	return {
-		mechanism: wordpressAuth.mechanism,
-		publicTokenField:
-			wordpressAuth.publicTokenField ??
-			WP_TYPIA_OPENAPI_LITERALS.PUBLIC_WRITE_TOKEN_FIELD,
-	};
-}
-
-function deriveLegacyAuthModeFromNormalizedAuth(
-	auth: EndpointAuthIntent,
-	wordpressAuth: EndpointWordPressAuthDefinition | undefined,
-): EndpointOpenApiAuthMode | undefined {
-	if (auth === "public") {
-		return "public-read";
-	}
-
-	if (
-		auth === "authenticated" &&
-		wordpressAuth?.mechanism ===
-			WP_TYPIA_OPENAPI_LITERALS.WORDPRESS_REST_NONCE_MECHANISM
-	) {
-		return "authenticated-rest-nonce";
-	}
-
-	if (
-		auth === "public-write-protected" &&
-		wordpressAuth?.mechanism ===
-			WP_TYPIA_OPENAPI_LITERALS.WORDPRESS_PUBLIC_TOKEN_MECHANISM
-	) {
-		return "public-signed-token";
-	}
-
-	return undefined;
-}
-
-function compareNormalizedEndpointAuth(
-	left: NormalizedEndpointAuthDefinition,
-	right: NormalizedEndpointAuthDefinition,
-): boolean {
-	return (
-		left.auth === right.auth &&
-		left.authMode === right.authMode &&
-		left.wordpressAuth?.mechanism === right.wordpressAuth?.mechanism &&
-		left.wordpressAuth?.publicTokenField === right.wordpressAuth?.publicTokenField
-	);
-}
-
-/**
- * Normalizes endpoint auth metadata into backend-neutral intent plus optional
- * WordPress adapter details.
- *
- * This public runtime helper accepts either the authored `auth` and optional
- * `wordpressAuth` shape or the deprecated `authMode` field. It validates that
- * the provided fields are compatible, resolves the legacy compatibility mode,
- * and returns a normalized definition without mutating the input endpoint.
- *
- * @param endpoint - Endpoint auth fields to normalize, including `auth`,
- * `authMode`, `wordpressAuth`, and optional identity fields used in error
- * messages (`operationId`, `path`, and `method`).
- * @returns The normalized auth definition for the endpoint.
- * @throws When `auth` and deprecated `authMode` conflict.
- * @throws When `wordpressAuth` is attached to the `public` auth intent.
- * @throws When the selected `wordpressAuth` mechanism is incompatible with the
- * chosen auth intent.
- * @throws When neither `auth` nor deprecated `authMode` is defined.
- */
-export function normalizeEndpointAuthDefinition(
-	endpoint: Pick<
-		EndpointOpenApiEndpointDefinition,
-		"auth" | "authMode" | "operationId" | "path" | "wordpressAuth" | "method"
-	>,
-): NormalizedEndpointAuthDefinition {
-	const endpointDescription =
-		typeof endpoint.operationId === "string" &&
-		typeof endpoint.path === "string" &&
-		typeof endpoint.method === "string"
-			? formatEndpointDescription(endpoint as EndpointOpenApiEndpointDefinition)
-			: "the current endpoint";
-	const nextWordPressAuth = normalizeWordPressAuthDefinition(
-		endpoint.wordpressAuth,
-	);
-
-	let normalized: NormalizedEndpointAuthDefinition | null = null;
-	if (endpoint.auth) {
-		normalized = {
-			auth: endpoint.auth,
-			authMode: deriveLegacyAuthModeFromNormalizedAuth(
-				endpoint.auth,
-				nextWordPressAuth,
-			),
-			...(nextWordPressAuth ? { wordpressAuth: nextWordPressAuth } : {}),
-		};
-
-		if (endpoint.auth === "public" && nextWordPressAuth) {
-			throw new Error(
-				`Endpoint "${endpointDescription}" cannot attach wordpressAuth when auth intent is "public".`,
-			);
-		}
-		if (
-			endpoint.auth === "authenticated" &&
-			nextWordPressAuth?.mechanism ===
-				WP_TYPIA_OPENAPI_LITERALS.WORDPRESS_PUBLIC_TOKEN_MECHANISM
-		) {
-			throw new Error(
-				`Endpoint "${endpointDescription}" uses auth intent "authenticated" but wordpressAuth mechanism "${nextWordPressAuth.mechanism}" only supports "public-write-protected".`,
-			);
-		}
-		if (
-			endpoint.auth === "public-write-protected" &&
-			nextWordPressAuth?.mechanism ===
-				WP_TYPIA_OPENAPI_LITERALS.WORDPRESS_REST_NONCE_MECHANISM
-		) {
-			throw new Error(
-				`Endpoint "${endpointDescription}" uses auth intent "public-write-protected" but wordpressAuth mechanism "${nextWordPressAuth.mechanism}" only supports "authenticated".`,
-			);
-		}
-	}
-
-	let legacyNormalized: NormalizedEndpointAuthDefinition | null = null;
-	if (endpoint.authMode) {
-		legacyNormalized =
-			endpoint.authMode === "public-read"
-				? {
-						auth: "public",
-						authMode: "public-read",
-				  }
-				: endpoint.authMode === "authenticated-rest-nonce"
-					? {
-							auth: "authenticated",
-							authMode: "authenticated-rest-nonce",
-							wordpressAuth: {
-								mechanism:
-									WP_TYPIA_OPENAPI_LITERALS.WORDPRESS_REST_NONCE_MECHANISM,
-							},
-					  }
-					: {
-							auth: "public-write-protected",
-							authMode: "public-signed-token",
-							wordpressAuth: {
-								mechanism:
-									WP_TYPIA_OPENAPI_LITERALS.WORDPRESS_PUBLIC_TOKEN_MECHANISM,
-								publicTokenField:
-									WP_TYPIA_OPENAPI_LITERALS.PUBLIC_WRITE_TOKEN_FIELD,
-							},
-					  };
-	}
-
-	if (normalized && legacyNormalized) {
-		if (!compareNormalizedEndpointAuth(normalized, legacyNormalized)) {
-			throw new Error(
-				`Endpoint "${endpointDescription}" defines conflicting auth metadata between auth/wordpressAuth and deprecated authMode.`,
-			);
-		}
-
-		return normalized;
-	}
-
-	if (normalized) {
-		return normalized;
-	}
-
-	if (legacyNormalized) {
-		return legacyNormalized;
-	}
-
-	throw new Error(
-		`Endpoint "${endpointDescription}" must define either auth or deprecated authMode.`,
-	);
-}
-
 function getContractSchemaName(
 	contractKey: string,
 	contract: EndpointOpenApiContractDocument | undefined,
@@ -1133,44 +576,6 @@ function buildQueryParameters(contract: EndpointOpenApiContractDocument): OpenAp
 		required: attribute.ts.required !== false,
 		schema: manifestAttributeToJsonSchema(attribute),
 	}));
-}
-
-function createBootstrapResponseHeaders(
-	normalizedAuth: NormalizedEndpointAuthDefinition,
-): Record<string, JsonSchemaObject> {
-	const headers: Record<string, JsonSchemaObject> = {
-		"Cache-Control": {
-			description:
-				"Must be non-cacheable for fresh bootstrap write/session state.",
-			schema: {
-				type: "string",
-				example: "private, no-store, no-cache, must-revalidate",
-			},
-		},
-		Pragma: {
-			description: "Legacy non-cacheable bootstrap response directive.",
-			schema: {
-				type: "string",
-				example: "no-cache",
-			},
-		},
-	};
-
-	if (
-		normalizedAuth.wordpressAuth?.mechanism ===
-		WP_TYPIA_OPENAPI_LITERALS.WORDPRESS_REST_NONCE_MECHANISM
-	) {
-		headers.Vary = {
-			description:
-				"Viewer-aware bootstrap responses should vary on cookie-backed auth state.",
-			schema: {
-				type: "string",
-				example: "Cookie",
-			},
-		};
-	}
-
-	return headers;
 }
 
 function createSuccessResponse(
